feat(app-platform): Delete External Issue UI
Implements the front-end functionality necessary to delete External
Issues created through Sentry Apps.

Also fixes a problem where the "Link" tab of the new new Issue modal
wasn't submitting the correct data. It was relying on `initialData` to
set static things like `action`, but that wasn't being udpated when the
User navigated around the modal.

Instead, `SentryAppExternalIssueForm` keeps track of it's own
`FormModel` and updates fields on that directly. This was the only way I
could figure out how to pass data that isn't to be rendered, to the
form. Ideally I think this would be done via hidden input fields.